+++ lib/cargoLlvmCov.nix
@@ -1,0 +1,31 @@
+{ mkCargoDerivation
+, cargo-llvm-cov
+}:
+
+{ cargoExtraArgs ? ""
+, cargoLlvmCovCommand ? "test"
+, cargoLlvmCovExtraArgs ? "--lcov --output-path $out"
+, ...
+}@origArgs:
+
+let
+  args = builtins.removeAttrs origArgs [
+    "cargoExtraArgs"
+    "cargoLlvmCovCommand"
+    "cargoLlvmCovExtraArgs"
+  ];
+in
+
+mkCargoDerivation (args // {
+  pnameSuffix = "-llvm-cov";
+
+  nativeBuildInputs = (args.nativeBuildInputs or [ ]) ++ [ cargo-llvm-cov ];
+
+  doInstallCargoArtifacts = false;
+
+  buildPhaseCargoCommand = ''
+    cargoWithProfile llvm-cov "${cargoLlvmCovCommand}" ${cargoExtraArgs} ${cargoLlvmCovExtraArgs}
+  '';
+
+  installPhaseCommand = "";
+})