+++ lib/setupHooks/cargoHelperFunctionsHook.sh
@@ -10,5 +10,11 @@ cargo() {
 # Injects `--profile $CARGO_PROFILE` into a particular cargo invocation
 # if the environment variable is set
 cargoWithProfile() {
-  cargo "${@:1:1}" ${CARGO_PROFILE:+--profile ${CARGO_PROFILE}} "${@:2}"
+  local profileArgs
+  if [[ "${CARGO_PROFILE}" == "release" ]]; then
+    profileArgs="--release"
+  else
+    profileArgs="${CARGO_PROFILE:+--profile ${CARGO_PROFILE}}"
+  fi
+  cargo "${@:1:1}" ${profileArgs} "${@:2}"
 }